fix: sweep cycle — 9 bugs + 29 kernel __shfl_xor_sync fixes (#954-#962)
* fix: AMD HIP compatibility — NaN-free inference on Strix Halo

## Root cause
fused_qkv_kernel and 42 other kernels use deprecated __shfl_xor warp
shuffle without _sync mask. On AMD HIP (gfx1151), this produces garbage
NaN values in QKV output, propagating through attention/MoE to argmax.

## Fixes
- Replace __shfl_xor → __shfl_xor_sync(0xffffffffULL, ...) in 42 .hip files
- Replace fused_qkv_kernel with mm_k (no warp shuffle, AMD compatible)
- Add f16_to_f32() correct IEEE 754 half→float conversion
- Clip FP16 weights to ±32 during loading (prevent 2048-dim overflow)
- Clamp mm_k output to ±65504 (prevent __float2half inf)
- Fix O-proj in-place buffer race (d_ao→d_ao → d_tmp→d_ao)
- Disable HIP graph capture (stream shared with NPU pipeline)
- Add nan_clean_k after QKV and attention residual
- Fix trailing-slash in weights path
- Deduplicate .bin model discovery

## Results
- ZAYA1-8B produces real tokens (no NaN) on Strix Halo ROCm HIP
- 6.3 tok/s with mm_k, MoE enabled
- zaya_full benchmark: 40 layers in 933ms, correct output
- RMSNorm-only path: 128 tok/s

## Known limitations
- nan_clean on d_hs prevents NaN but limits attention context → repetitive output
- mm_k is 4x slower than fused_qkv_kernel (fused produces NaN on AMD)
- __shfl_xor_sync blanket fix may need per-kernel mask tuning


* fix: Zamba2 attention_forward only processed 1 of 32 heads — 97% garbage output (#946)

Root cause: attention_forward() iterated n_groups (n_heads/n_kv_heads)
instead of n_heads. With the default config (32 heads, 32 KV heads),
n_groups=1, so only QUERY HEAD 0 was computed. The remaining 31 heads
(97% of the attention output) were left uninitialized, producing the
incoherent/garbage output reported in #946.

Also fixed the attn_head_dim fallback in the GGUF loader:
- Old: d_inner/n_attn_heads = 5120/32 = 160 (WRONG)
- New: derive from actual attn_q tensor shape, or default to 80

Changes:
1. attention_forward(): iterate over ALL query heads (h=0..n_heads-1),
   with GQA mapping h -> KV head = h / n_groups
2. head_dim fallback: try rope.dimension_count first, then derive from
   attn_q tensor shape, then default to 80

* fix: add server-side generation timeout + RLIMIT_AS OOM safety net (#948)

Implements two safety mechanisms to prevent a single slow/memory-hungry
request from OOM-killing the entire server:

1. Generation timeout:
   - New global g_generation_timeout_ms (default: 600000ms = 10 min)
   - Configurable via CLI --gen-timeout-ms or GEN_TIMEOUT_MS env var
   - Checked per-token in generate_completion() — when exceeded,
     the loop breaks, g_inference_mutex is released promptly, and
     partial results are returned with finish_reason: "timeout"
   - Also checked during prefill loop
   - Exposed in /v1/health as generation_timeout_ms

2. RLIMIT_AS memory ceiling:
   - Set at startup to 256 GB (safety net for 122 GB physical RAM host)
   - malloc returns NULL (or triggers graceful error) instead of OOM
   - Ensures the kernel OOM killer can't take down the process

This prevents the scenario from issue #948 where a zamba-7b-v1 request
grew to 56.5GB over ~11 minutes, got OOM-killed, and took the whole
singleton inference server down.

* fix: MoE tensor shape validation + architecture guard for generic GGUF backend (#947)

Two safety fixes for the generic CPU GGUF backend to prevent loading
models with incompatible tensor layouts:

1. load_tensor() now returns SIZE_MAX on shape mismatch (not 0):
   - 0 is a valid flat_weights index, so a mismatched tensor silently
     pointed to the first loaded tensor — wrapping ZAYA1 MoE tensor
     shapes into invalid indices that produced degenerate output
   - Now correctly returns SIZE_MAX like load_tensor_optional()

2. Layer loading aborts early on any shape mismatch:
   - Previously, a single failed tensor (e.g. ffn_gate_inp) returned 0,
     and the model continued with garbage data
   - Now checks all required tensors per-layer and returns false from
     load_gguf() if any shape is wrong

3. Architecture guard rejects unsupported architectures:
   - RCPP_ARCH_ZAYA, RCPP_ARCH_ZAMBA2, RCPP_ARCH_ZAMBA, RCPP_ARCH_MAMBA
     all have their own dedicated backends with different tensor layouts
   - The generic backend would previously try to load them, fail silently
     on shape checks (returning 0 indices), then produce garbage
   - Now cleanly refuses with a diagnostic message

4. w() accessor bounds-checks SIZE_MAX indices:
   - Returns a pointer to a static zero instead of flat_weights[SIZE_MAX]

Closes #947

* fix: sweep cycle — 9 bugs + 29 kernel __shfl_xor_sync fixes (#954-#962)

---------

+++ src/zamba2_engine.h
@@ -184,11 +184,14 @@ struct Zamba2Model {
     // Initialize state buffers
     bool init_state() {
         int64_t conv_dim = cfg.d_inner + 2 * cfg.n_group * cfg.d_state;
-        conv_states.resize(cfg.n_layers * (cfg.d_conv - 1) * conv_dim, 0.0f);
-        ssm_states.resize(cfg.n_layers * cfg.d_state * cfg.d_inner, 0.0f);
+        // Use size_t cast for each term to prevent 32-bit int overflow
+        // in multiplication (issue #962). n_layers * d_state * d_inner can
+        // exceed INT_MAX for large models.
+        conv_states.resize((size_t)cfg.n_layers * (cfg.d_conv - 1) * conv_dim, 0.0f);
+        ssm_states.resize((size_t)cfg.n_layers * cfg.d_state * cfg.d_inner, 0.0f);
         // Allocate KV cache: one slot per hybrid layer
         int n_hybrid_layers = num_hybrid_layers();
-        kv_cache.resize(n_hybrid_layers * 2 * cfg.max_seq_len * cfg.n_kv_heads * cfg.attn_head_dim, 0.0f);
+        kv_cache.resize((size_t)n_hybrid_layers * 2 * cfg.max_seq_len * cfg.n_kv_heads * cfg.attn_head_dim, 0.0f);
         pos = 0;
         return true;
     }
